Fixed the model creation inside DBContext

+++ src/BoligInfo.Core/Models/Loan.cs
@@ -4,10 +4,10 @@ namespace BoligInfo.Core.Models;
 
 public class Loan
 {
-    public int Id { get; set; }
+    public long Id { get; set; }
     public LoanType LoanType { get; set; }
     public double LoanAmount { get; set; }
     public double InterestRate { get; set; }
     public int LoanLifetime { get; set; }
-    public int EquityId { get; set; }
+    public long EquityId { get; set; }
 }
